+++ docs/CHANGELOG.md
@@ -49,6 +49,7 @@
     * `3D Dropped Items` (Enabled By Default)
     * `Render Entity Shadows` (Enabled By Default)
     * `Render Vignette` (Enabled By Default)
+    * `Increase Render Chunk Size` (Enabled By Default)
   * Existing Functionality (All Enabled By Default)
     * `Fix Screen Rendering When Hiding HUD`
     * `Sanitize Usernames`